Refactor place container to use redux
