+++ client/components/SavedPlaceContainer.js
@@ -1,6 +1,5 @@
 import React, { PropTypes, Component } from 'react';
 import { connect } from 'react-redux';
-
 import SavedPlaceEntry from './SavedPlaceEntry.js';
 
 class SavedPlaceContainer extends Component {
